Restructured
FuzzyNumber: split from Number;
Number: apply(factor) now creates unit value.
