+++ src/main/scala/com/phasmidsoftware/number/core/FuzzyNumber.scala
@@ -1,0 +1,46 @@
+package com.phasmidsoftware.number.core
+
+/**
+  * This class is designed to model a fuzzy Numeral.
+  * See Number for more details on the actual representation.
+  *
+  * @param value  the value of the Number, expressed as a nested Either type.
+  * @param factor the scale factor of the Number: valid scales are: Scalar, Pi, and E.
+  * @param fuzz   the fuzziness of this Number.
+  */
+case class FuzzyNumber(override val value: Value, override val factor: Factor, fuzz: Option[Fuzz[Double]]) extends Number(value, factor) with Fuzzy[Double] {
+
+  /**
+    * Auxiliary constructor for an exact number.
+    *
+    * @param v    the value for the new Number.
+    * @param fuzz the fuzz for the new Number.
+    */
+  def this(v: Value, fuzz: Option[Fuzz[Double]]) = this(v, Scalar, fuzz)
+
+  /**
+    * Make a copy of this Number, given the same degree of fuzziness as the original.
+    * Both the value and the factor will be changed.
+    *
+    * @param v the value.
+    * @param f the factor.
+    * @return a FuzzyNumber.
+    */
+  protected def makeNumber(v: Value, f: Factor): Number = makeNumber(v, f, fuzz)
+
+  /**
+    * Make a copy of this Number, given the same degree of fuzziness as the original.
+    * Both the value and the factor will be changed.
+    * CONSIDER: not entirely sure we need this method.
+    *
+    * @param v the value.
+    * @param f the factor.
+    * @param z the new fuzziness.
+    * @return a FuzzyNumber.
+    */
+  protected def makeNumber(v: Value, f: Factor, z: Option[Fuzz[Double]]): Number = FuzzyNumber(v, f, z)
+}
+
+object FuzzyNumber {
+  def apply(): Number = Number.apply()
+}
